Try 4 frames
The issue I've been hunting appears to be a firmware error in the camera. In H.264 encoding mode it will "pause" every 65,534 frames _exactly_ - even if the camera is not currently connected (meaning if I stop capturing for a few minutes, I can add that duration of frames to predict the next drop). So...hoping 4 frames will cover the delay...